agenttel-python/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "agenttel"
version = "0.1.0"
description = "AgentTel Python SDK — AI-native telemetry for Python services"
readme = "README.md"
license = "Apache-2.0"
requires-python = ">=3.11"
authors = [{ name = "AgentTel", email = "engineering@agenttel.io" }]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: System :: Monitoring",
    "Framework :: FastAPI",
]
dependencies = [
    "opentelemetry-api>=1.20.0",
    "opentelemetry-sdk>=1.20.0",
    "pydantic>=2.0",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
fastapi = ["fastapi>=0.100.0", "opentelemetry-instrumentation-fastapi>=0.41b0"]
openai = ["openai>=1.0"]
anthropic = ["anthropic>=0.20"]
langchain = ["langchain-core>=0.1"]
bedrock = ["boto3>=1.28"]
agent = ["aiohttp>=3.9"]
all = ["agenttel[fastapi,openai,anthropic,langchain,bedrock,agent]"]
dev = [
    "pytest>=7.0",
    "pytest-asyncio>=0.21",
    "httpx>=0.24",
    "respx>=0.20",
]

[project.entry-points."opentelemetry_traces"]
agenttel = "agenttel:AgentTelSpanProcessor"

[tool.hatch.build.targets.wheel]
packages = ["src/agenttel"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
